Cleaned up Merchant class, separated MerchantRepository
