+++ lib/merchant.rb
@@ -1,31 +1,6 @@
-require 'pry'
-require 'csv'
+require './lib/merchant_repository'
 
-class FakeRepository
-  attr_reader  :all
-
-  def initialize
-    @handle = CSV.open './data/merchants.csv', headers: true, header_converters: :symbol
-    @all    = @handle.map { |line| Merchant.new(line).merchant }
-  end
-
-  def find_by_id(id)
-    @all.find { |merchant| merchant["id"].eql?(id) }
-  end 
-
-  def find_by_name(name)
-    @all.find { |merchant| merchant["name"].eql?(name) }
-  end
-
-  def find_all_by_name(fragment)
-    results = @all.find_all do |merchant|
-      merchant["name"].downcase.include?(fragment.downcase)
-    end
-  results
-  end
-end
-
-class Merchant < FakeRepository
+class Merchant < MerchantRepository
   attr_reader  :merchant
 
   def initialize(line)
@@ -43,4 +18,4 @@ class Merchant < FakeRepository
 
 end
 
-puts FakeRepository.new.find_all_by_name("Jai")
+puts MerchantRepository.new.find_all_by_name("Jai")
